Split simulation part into its own module
